newlinter: renaming gofuncor to funcorder
