+++ pkg/golinters/funcorder/funcorder.go
@@ -1,7 +1,7 @@
-package gofuncor
+package funcorder
 
 import (
-	"github.com/manuelarte/gofuncor/pkg/analyzer"
+	"github.com/manuelarte/funcorder/pkg/analyzer"
 	"golang.org/x/tools/go/analysis"
 
 	"github.com/golangci/golangci-lint/v2/pkg/goanalysis"
