+++ pkg/lint/lintersdb/builder_linter.go
@@ -29,6 +29,7 @@ import (
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/fatcontext"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/forbidigo"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/forcetypeassert"
+	"github.com/golangci/golangci-lint/v2/pkg/golinters/funcorder"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/funlen"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/gci"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/ginkgolinter"
@@ -44,7 +45,6 @@ import (
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/godox"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/gofmt"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/gofumpt"
-	"github.com/golangci/golangci-lint/v2/pkg/golinters/gofuncor"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/goheader"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/goimports"
 	"github.com/golangci/golangci-lint/v2/pkg/golinters/golines"
@@ -255,6 +255,10 @@ func (LinterBuilder) Build(cfg *config.Config) ([]*linter.Config, error) {
 			WithLoadForGoAnalysis().
 			WithURL("https://github.com/gostaticanalysis/forcetypeassert"),
 
+		linter.NewConfig(funcorder.New()).
+			WithSince("v2.1.0").
+			WithURL("https://github.com/manuelarte/funcorder"),
+
 		linter.NewConfig(fatcontext.New(&cfg.Linters.Settings.Fatcontext)).
 			WithSince("v1.58.0").
 			WithLoadForGoAnalysis().
@@ -335,10 +339,6 @@ func (LinterBuilder) Build(cfg *config.Config) ([]*linter.Config, error) {
 			WithSince("v1.28.0").
 			WithAutoFix().
 			WithURL("https://github.com/mvdan/gofumpt"),
-
-		linter.NewConfig(gofuncor.New()).
-			WithSince("v2.1.0").
-			WithURL("https://github.com/manuelarte/gofuncor"),
 
 		linter.NewConfig(golines.New(&cfg.Linters.Settings.GoLines)).
 			WithSince("v2.0.0").
